+++ R/zzz.R
@@ -1,5 +1,7 @@
 # Set all default options on package load
 .onLoad <- function(libname, pkgname) {
-  set_logfile(confirm = FALSE, create = FALSE)
-  set_timestamp_format(confirm = FALSE)
+  setup_echo()
+  setup_log_level()
+  setup_logfile()
+  setup_timestamp_format()
 }